+++ src/winimg.h
@@ -17,7 +17,8 @@ extern void (winimgs_paint)(struct term* term_p);
 #define winimgs_clear(...) (winimgs_clear)(term_p, ##__VA_ARGS__)
 extern void (winimgs_clear)(struct term* term_p);
 
-extern void win_emoji_show(int x, int y, wchar * efn, void * * bufpoi, int * buflen, int elen, ushort lattr, bool italic);
+#define win_emoji_show(...) (win_emoji_show)(term_p, ##__VA_ARGS__)
+extern void (win_emoji_show)(struct term* term_p, int x, int y, wchar * efn, void * * bufpoi, int * buflen, int elen, ushort lattr, bool italic);
 
 extern void save_img(HDC, int x, int y, int w, int h, wstring fn);